Fix ensure_alloca() moar
The last commit had the logic for stack growth direction detection
reversed.

More importantly, it's important to allocate one page at a time,
repeatedly, rather than allocating the entire allocation in one go and
then test each page: if there's a signal handler for SIGSEGFAULT and
it's not on an alternate stack, then it the fault handling will scribble
over wherever the sp was set to.  We don't mind taking a SEGFAULT in
ensure_alloca(), but we do mind writing over anything beyond the guard
page!  Hat tip: Viktor Dukhovni, who noticed that Andy Polyakov noticed
this.

Both versions of ensure_alloca() included, the one that uses alloca()
and the one that uses VLAs, were checked against all the compilers at
http://gcc.godbolt.org/.  See http://goo.gl/LtJP9F